requirements.txt
Pillow==10.3.0
psutil==5.9.7
PyAutoGUI==0.9.54
pytesseract==0.3.10
